[Untested] Refactoring Pulsoid-specific code to allow for other providers
