+++ Providers/IHeartRateProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace HeartSpeak.Providers;
+
+public interface IHeartRateProvider
+{
+    public Task<int?> GetCurrentHeartRateAsync();
+}
